runs a pipeline
