+++ pipeline/training_haystack/.Dockerignore
@@ -1,0 +1,8 @@
+Dockerfile
+README.md
+*.pyc
+*.pyo
+*.pyd
+__pycache__
+__pycache__/
+Makefile
