+++ training_spec.json
@@ -1,0 +1,84 @@
+{
+  "pipelineSpec": {
+    "components": {
+      "comp-training": {
+        "executorLabel": "exec-training",
+        "inputDefinitions": {
+          "parameters": {
+            "bucket_name": {
+              "type": "STRING"
+            },
+            "data_path": {
+              "type": "STRING"
+            }
+          }
+        }
+      }
+    },
+    "deploymentSpec": {
+      "executors": {
+        "exec-training": {
+          "container": {
+            "args": [
+              "--executor_input",
+              "{{$}}",
+              "--function_to_execute",
+              "training"
+            ],
+            "command": [
+              "sh",
+              "-c",
+              "\nif ! [ -x \"$(command -v pip)\" ]; then\n    python3 -m ensurepip || python3 -m ensurepip --user || apt-get install python3-pip\nfi\n\nPIP_DISABLE_PIP_VERSION_CHECK=1 python3 -m pip install --quiet     --no-warn-script-location 'kfp==1.8.14' && \"$0\" \"$@\"\n",
+              "sh",
+              "-ec",
+              "program_path=$(mktemp -d)\nprintf \"%s\" \"$0\" > \"$program_path/ephemeral_component.py\"\npython3 -m kfp.v2.components.executor_main                         --component_module_path                         \"$program_path/ephemeral_component.py\"                         \"$@\"\n",
+              "\nimport kfp\nfrom kfp.v2 import dsl\nfrom kfp.v2.dsl import *\nfrom typing import *\n\ndef training(\n    bucket_name: str =\"filipegracio-haystack\",\n    data_path: str =\"data/myth/\"\n    ):\n    from haystack.utils import convert_files_to_docs, launch_es\n    from haystack.document_stores import FAISSDocumentStore\n    from haystack.nodes import PreProcessor, EmbeddingRetriever, FARMReader\n    from haystack.pipelines import ExtractiveQAPipeline\n\n    download_files(bucket_name=bucket_name, prefix=data_path, dl_dir=data_path)\n\n    launch_es()\n\n    all_docs = convert_files_to_docs(dir_path=data_path)\n\n\n    preprocessor = PreProcessor(\n        clean_empty_lines=True,\n        clean_whitespace=True,\n        clean_header_footer=False,\n        split_by=\"word\",\n        split_length=100,\n        split_respect_sentence_boundary=True,\n    )\n    docs = preprocessor.process(all_docs)\n\n    print(f\"n_files_input: {len(all_docs)}\\nn_docs_output: {len(docs)}\")\n\n    document_store = FAISSDocumentStore(faiss_index_factory_str=\"Flat\",  similarity=\"cosine\")\n\n\n    retriever = EmbeddingRetriever(\n        document_store=document_store,\n        embedding_model=\"sentence-transformers/multi-qa-mpnet-base-dot-v1\",\n        model_format=\"sentence_transformers\",\n    )\n\n\n    document_store.update_embeddings(retriever=retriever)\n\n    reader = FARMReader(model_name_or_path=\"deepset/roberta-base-squad2\", use_gpu=True)\n\n\n    extractive_pipe = ExtractiveQAPipeline(reader, retriever)\n\n    extractive_pipe.save_to_yaml(path='pipe.yaml')\n    document_store.save(\"my_faiss_index.faiss\")\n\n"
+            ],
+            "image": "europe-docker.pkg.dev/filipegracio-ai-learning/haystack-docker-repo/haystack-training:tag1"
+          }
+        }
+      }
+    },
+    "pipelineInfo": {
+      "name": "custom-model-training-sample-pipeline"
+    },
+    "root": {
+      "dag": {
+        "tasks": {
+          "training": {
+            "cachingOptions": {
+              "enableCache": true
+            },
+            "componentRef": {
+              "name": "comp-training"
+            },
+            "inputs": {
+              "parameters": {
+                "bucket_name": {
+                  "runtimeValue": {
+                    "constantValue": {
+                      "stringValue": "filipegracio-haystack"
+                    }
+                  }
+                },
+                "data_path": {
+                  "runtimeValue": {
+                    "constantValue": {
+                      "stringValue": "data/myth/"
+                    }
+                  }
+                }
+              }
+            },
+            "taskInfo": {
+              "name": "training"
+            }
+          }
+        }
+      }
+    },
+    "schemaVersion": "2.0.0",
+    "sdkVersion": "kfp-1.8.14"
+  },
+  "runtimeConfig": {}
+}
